requirements.txt
datasets>=2.0.0
google-genai>=0.5.0
